hotplace rev.842

+++ test/quic/test_quic_frame.cpp
@@ -12,11 +12,11 @@
 void test_ack() {
     _test_case.begin("ACK");
 
-    auto lambda = [](const char* func, const char* text, t_ovl_points<uint64>& p, ack_t& e) -> void {
+    auto lambda = [](const char* func, const char* text, t_ovl_points<uint32>& p, ack_t& e) -> void {
         ack_t ack;
         ack << p;
 
-        t_ovl_points<uint64> temp;
+        t_ovl_points<uint32> temp;
         ack >> temp;
 
         _test_case.assert(ack == e, func, text);
@@ -25,7 +25,7 @@ void test_ack() {
 
     {
         // #35 ACK(12, FAR:5)
-        t_ovl_points<uint64> part;
+        t_ovl_points<uint32> part;
         part.add(7).add(8).add(9).add(10).add(11).add(12);
 
         ack_t expect(12, 5);
@@ -34,7 +34,7 @@ void test_ack() {
     }
     {
         // #37 ACK(14, FAR:0, [0]G:0,R:5)
-        t_ovl_points<uint64> part;
+        t_ovl_points<uint32> part;
         part.add(7, 12).add(14);
 
         ack_t expect(14, 0);
@@ -44,7 +44,7 @@ void test_ack() {
     }
     {
         // #46 ACK(16, FAR:2, [0]G:0,R:5)
-        t_ovl_points<uint64> part;
+        t_ovl_points<uint32> part;
         part.add(7, 12).add(14).add(15, 16);
 
         ack_t expect(16, 2);
@@ -54,7 +54,7 @@ void test_ack() {
     }
     {
         // #47 ACK(18, FAR:4, [0]G:0,R:5)
-        t_ovl_points<uint64> part;
+        t_ovl_points<uint32> part;
         part.add(7, 12).add(14).add(15, 16).add(17, 18);
 
         ack_t expect(18, 4);
@@ -64,7 +64,7 @@ void test_ack() {
     }
     {
         // #48 ACK(21, FAR:0, [0]G:1,R:4, [1]G:0,R:5)
-        t_ovl_points<uint64> part;
+        t_ovl_points<uint32> part;
         part.add(7, 12).add(14).add(15, 16).add(17, 18).add(21);
 
         ack_t expect(21, 0);
@@ -75,7 +75,7 @@ void test_ack() {
     }
     {
         // #49 ACK(21, FAR:0, [0]G:0,R:5, [1]G:0,R:5)
-        t_ovl_points<uint64> part;
+        t_ovl_points<uint32> part;
         part.add(7, 12).add(14).add(15, 16).add(17, 18).add(21).add(19);
 
         ack_t expect(21, 0);
@@ -86,7 +86,7 @@ void test_ack() {
     }
     {
         // #50 ACK(22, FAR:8, [0]G:0,R:5)
-        t_ovl_points<uint64> part;
+        t_ovl_points<uint32> part;
         part.add(7, 12).add(14).add(15, 16).add(17, 18).add(21).add(19).add(22).add(20);
 
         ack_t expect(22, 8);
